Improve Copyright year parsing
According to the GPX schema, the copyright year is an xsd:gYear which
can have an optional timezone suffix.

Handle this with an explicit UnmarshalXML implementation for
CopyrightType. It's still not a full xsd:gYear parser, as it can't
handle negative dates, but it's unlikely there'll be a negative
copyright year.

This is necessary to parse GPX files exported from Openrouteservice